Move Update classes into their own directory
