--- src/Objects/EditedMessage.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Telegram\Bot\Objects;
-
-/**
- * Class EditedMessage
- */
-class EditedMessage extends Message
-{
-}
