+++ src/Objects/Updates/PollAnswer.php
@@ -1,6 +1,9 @@
 <?php
 
-namespace Telegram\Bot\Objects;
+namespace Telegram\Bot\Objects\Updates;
+
+use Telegram\Bot\Objects\BaseObject;
+use Telegram\Bot\Objects\User;
 
 /**
  * Class PollAnswer.
