+++ src/Objects/Updates/PreCheckoutQuery.php
@@ -1,11 +1,14 @@
 <?php
 
-namespace Telegram\Bot\Objects\Payments;
+namespace Telegram\Bot\Objects\Updates;
 
 use Telegram\Bot\Objects\BaseObject;
+use Telegram\Bot\Objects\Payments\OrderInfo;
 use Telegram\Bot\Objects\User;
 
 /**
+ * Class PreCheckoutQuery
+ *
  * @link https://core.telegram.org/bots/api#precheckoutquery
  *
  * @property string    $id                       Unique query identifier
